회원가입 페이지 생성 로그인 비밀번호 업데이트 회원가입시 SHA256 추가 회원가입 성공 시 자동 로그인 추가

+++ src/components/user/UserDetail.vue
@@ -31,7 +31,7 @@
 			</div>
 		</section>
 		<section>
-			<router-link to="/user/update">
+			<router-link :to="`/user/update`">
 				<button>비밀번호 수정</button>
 			</router-link>
 			<button @click="deleteUser">회원탈퇴</button>
@@ -53,10 +53,16 @@ export default {
 	},
 	methods: {
 		userLogout() {
-			this.$store.commit("DELETE_TOKEN");
+			if (confirm(`로그아웃 하시겠습니까??`))
+				this.$store.commit("DELETE_TOKEN");
 		},
 		deleteUser() {
-			this.$store.dispatch("FETCH_DELETE_USER", this.$store.state.user.id);
+			if (
+				confirm(
+					`WARNNING!!!\n회원 정보를 삭제하시겠습니까??\n해당 아이디와 관련된 모든 데이터가 영구적으로 삭제됩니다.`
+				)
+			)
+				this.$store.dispatch("FETCH_DELETE_USER", this.$store.state.user.id);
 		},
 	},
 };
